Safari-611.1.10/LayoutTests/accessibility/mac/invalid-status-for-input-types-expected.txt


This tests that we are exposing correct invalid status for different types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: AXInvalid is true.
PASS: AXInvalid is false.
PASS: AXInvalid is true.
PASS: AXInvalid is false.
PASS: AXInvalid is true.
PASS: AXInvalid is false.
